requirements.txt

unset __conda_setup
# <<< conda initialize <<<

PyQt5
pillow